@@ -7,10 +7,10 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack-dashboard -t 'React-Redux Boilerplate' -- webpack-dev-server --colors --no-info",
     "dev": "webpack-dev-server --colors",
-    "build": "NODE_ENV=\"production\" webpack",
+    "build": "cross-env NODE_ENV=\"production\" webpack",
     "lint-fix": "eslint ./src ./webpack.config.js -f table --ext .js --ext .jsx --fix || true",
     "lint": "eslint ./src ./webpack.config.js -f table --ext .js --ext .jsx || true",
-    "preview": "NODE_ENV=\"production\" ./node_modules/webpack-dashboard/bin/webpack-dashboard.js -t 'Preview Mode - React-Redux Boilerplate' -- ./node_modules/webpack-dev-server/bin/webpack-dev-server.js",
+    "preview": "cross-env NODE_ENV=\"production\" ./node_modules/webpack-dashboard/bin/webpack-dashboard.js -t 'Preview Mode - React-Redux Boilerplate' -- ./node_modules/webpack-dev-server/bin/webpack-dev-server.js",
     "hook-add": "prepush install",
     "hook-remove": "prepush remove"
   },
@@ -18,7 +18,7 @@
     "type": "git",
     "url": "git+https://github.com/zsxsoft/blog.zsxsoft.com.git"
   },
-  "author": "zsx<zsx@zsxsot.com>",
+  "author": "zsx<zsx@zsxsoft.com>",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/zsxsoft/blog.zsxsoft.com/issues"
@@ -26,56 +26,57 @@
   "homepage": "/",
   "dependencies": {
     "animated-scrollto": "^1.1.0",
-    "anujs": "^1.1.0",
+    "anujs": "^1.1.2",
     "classnames": "^2.2.5",
+    "cross-env": "^5.0.5",
     "dynamics.js": "^1.1.5",
     "fastclick": "^1.0.6",
     "lodash": "^4.17.4",
-    "material-ui": "^0.19.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "material-ui": "^0.19.3",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-flip-move": "^2.9.15",
-    "react-lite": "^0.15.38",
+    "react-lite": "^0.15.39",
     "react-router-dom": "^4.2.2",
-    "react-transition-group": "^1.2.0",
+    "react-transition-group": "^2.2.1",
     "social-share.js": "^1.0.16",
     "style-it": "^1.6.13",
     "trianglify": "^1.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.3",
+    "autoprefixer": "^7.1.5",
     "babel": "^6.23.0",
     "babel-core": "^6.26.0",
-    "babel-eslint": "^7.2.3",
+    "babel-eslint": "^8.0.1",
     "babel-loader": "^7.1.2",
     "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-react-optimize": "^1.0.1",
     "babel-preset-stage-0": "^6.24.1",
-    "css-loader": "0.28.5",
-    "duplicate-package-checker-webpack-plugin": "^1.2.5",
-    "eslint": "^4.5.0",
+    "css-loader": "0.28.7",
+    "duplicate-package-checker-webpack-plugin": "^1.2.6",
+    "eslint": "^4.8.0",
     "eslint-config-standard": "^10.2.1",
     "eslint-config-standard-react": "^5.0.0",
     "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.1.1",
+    "eslint-plugin-node": "^5.2.0",
     "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^7.3.0",
+    "eslint-plugin-react": "^7.4.0",
     "eslint-plugin-standard": "^3.0.1",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
+    "extract-text-webpack-plugin": "^3.0.1",
+    "file-loader": "^1.1.5",
     "html-webpack-plugin": "^2.30.1",
     "node-sass": "^4.5.3",
-    "offline-plugin": "^4.8.3",
+    "offline-plugin": "^4.8.4",
     "postcss-loader": "^2.0.6",
     "prepush": "^3.1.11",
     "react-hot-loader": "^3.0.0-beta.6",
     "sass-loader": "^6.0.6",
-    "style-loader": "^0.18.2",
-    "url-loader": "^0.5.9",
-    "webpack": "^3.5.5",
+    "style-loader": "^0.19.0",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.6.0",
     "webpack-bundle-analyzer": "^2.9.0",
     "webpack-dashboard": "^1.0.0-5",
-    "webpack-dev-server": "^2.7.1"
+    "webpack-dev-server": "^2.9.1"
   }
 }
