@@ -1,9 +1,10 @@
 import { List, ListItem, makeSelectable } from 'material-ui/List'
-import React, { PropTypes, PureComponent } from 'react'
+import React, { PureComponent } from 'react'
 import { spacing, typography, zIndex } from 'material-ui/styles'
 
 import Divider from 'material-ui/Divider'
 import Drawer from 'material-ui/Drawer'
+import PropTypes from 'prop-types'
 import { grey800 } from 'material-ui/styles/colors'
 
 const SelectableList = makeSelectable(List)
